[FEAT] 메인페이지 기록, 뱃지 조회 api 구현

+++ src/main/java/com/d204/rumeet/record/model/dto/RaceInfoDto.java
@@ -8,7 +8,7 @@ public class RaceInfoDto {
     int raceId;
     int userId;
     double velocity;
-    float km;
+    double km;
     long time;
     int heartRate;
     double kcal;
