+++ src/main/java/com/d204/rumeet/record/model/dto/RecordDto.java
@@ -11,7 +11,7 @@ public class RecordDto {
     int totalCount;
     double totalKm;
     long totalTime;
-    double averagePace;
+    int averagePace;
     int teamSuccessCount;
     int competitionSuccessCount;
